requirements.txt
numpy>=1.21.0
pandas>=1.3.0
rasterio>=1.2.0
geopandas>=0.9.0
scikit-image>=0.19.0
torch>=1.9.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.0
requests>=2.26.0
python-dotenv>=0.19.0